dsps_fluid firmware: Add alternate firmware from Samsung Skyrocket
This firmware is for a sensors subsystem (see CONFIG_MSM_DSPS in the
kernel).

I do not know what version these firmware blobs are, but they are
different from what we have now.

The current kernel actually doesn't load these blobs because
CONFIG_MSM_DSPS is not set in the kernel. It's likely that the dsps is
not used for anything on presto.